Reduced task interval for CheckForFinishedDownloadCommand (to every 10 seconds)
Migration enables the use of double intervals in minutes.
Also increased the scheduler executecommands interval to 1 second
instead of 30

+++ src/NzbDrone.Api/System/Tasks/TaskResource.cs
@@ -7,7 +7,7 @@ namespace NzbDrone.Api.System.Tasks
     {
         public String Name { get; set; }
         public String TaskName { get; set; }
-        public Int32 Interval { get; set; }
+        public double Interval { get; set; }
         public DateTime LastExecution { get; set; }
         public DateTime NextExecution { get; set; }
     }
